for real, up to step 5
Writing HTML fragments in JavaScript and programmatically inserting them
into the DOM is tedious. It makes your application harder to write and
harder to maintain. HTML templates address this issue by decoupling the
UI definition (HTML markup) from your code. There are a number of great
HTML template solutions: Mustache.js, Handlebar.js, and Underscore.js to
name a few.

In this section, we create two templates to streamline the code of the
Employee Directory application. We use Handlebar.js but the smae result
can be achieved using the other HTML template solutions.