Part 10 - Now is totally done!
In this section, we add the ability to tag an employee with his/her
location information. In this sample application, we display the raw
information (longitude/latitude) in the employee view. In a real-life
application, we would typically save the location in the database as
part of the employee information and show it on a map... and now it
works fine!